Back to using floats.

+++ PSXObj/Q3_12/Q3_12.cs
@@ -34,7 +34,7 @@ public class Q3_12
         value = v;
     }
 
-    public float ToFloat() => value / ONE;
+    public float ToFloat() => value / ((float)ONE);
     public override string ToString() => value.ToString();
 
     public override bool Equals(object? o)
